Several Bard fixes:
Bards will now receive messages as detrimental spells pulse. (If too many songs are spamming, please let me know which ones should and shouldn't - I've never played a Bard.)

Detrimental Bard songs will now display the yellow worn off message when the song wears off (unless the target died.)

Bard Charm will now be castable over Bard Mez. The new pet will be mezzed until that song wears off, and then function normally for the duration of the charm.

Bard group beneficial AEs no longer have a range check.
---

Fixed the SPELLS__STACKING log category. DO NOT enable this category unless you need to work with the stacking code. It creates HUGE logs!!!

Removed the newline tacked on the end of logs that were royally pissing me off. Added some more misc spell logging.

More changes to Knockback to handle pull, and for further accuracy.

+++ zone/command.cpp
@@ -10441,11 +10441,11 @@ void command_push(Client *c, const Seperator *sep){
 	else
 	{
 		bool success = false;
-		uint32 pushback = atoi(sep->arg[1]);
+		float pushback = atof(sep->arg[1]);
 		t = c->GetTarget();
 		if(sep->IsNumber(2))
 		{
-			uint32 pushup = atoi(sep->arg[2]);
+			float pushup = atof(sep->arg[2]);
 			if(t->DoKnockback(c, pushback, pushup))
 			{
 				success = true;
@@ -10461,7 +10461,7 @@ void command_push(Client *c, const Seperator *sep){
 
 		if(success)
 		{
-			c->Message(0, "%s was pushed for %i!", t->GetCleanName(), pushback);
+			c->Message(0, "%s was pushed for %0.1f!", t->GetCleanName(), pushback);
 			return;
 		}
 		else
